Supply Redis dsn on testing build as ArqRedis connects on construction
